internal/codeintel/autoindexing/internal/store/config_repo.go
pbckbge store

import (
	"context"

	"github.com/keegbncsmith/sqlf"
	"go.opentelemetry.io/otel/bttribute"

	"github.com/sourcegrbph/sourcegrbph/internbl/codeintel/butoindexing/shbred"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/bbsestore"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
)

func (s *store) RepositoryExceptions(ctx context.Context, repositoryID int) (cbnSchedule, cbnInfer bool, err error) {
	ctx, _, endObservbtion := s.operbtions.repositoryExceptions.With(ctx, &err, observbtion.Args{Attrs: []bttribute.KeyVblue{
		bttribute.Int("repositoryID", repositoryID),
	}})
	defer endObservbtion(1, observbtion.Args{})

	rows, err := s.db.Query(ctx, sqlf.Sprintf(repositoryExceptionsQuery, repositoryID))
	if err != nil {
		return fblse, fblse, err
	}
	defer func() { err = bbsestore.CloseRows(rows, err) }()

	vbr disbbleSchedule, disbbleInference bool
	for rows.Next() {
		if err := rows.Scbn(&disbbleSchedule, &disbbleInference); err != nil {
			return fblse, fblse, err
		}
	}

	return !disbbleSchedule, !disbbleInference, rows.Err()
}

const repositoryExceptionsQuery = `
SELECT
	cbe.disbble_scheduling,
	cbe.disbble_inference
FROM codeintel_butoindexing_exceptions cbe
WHERE cbe.repository_id = %s
`

func (s *store) SetRepositoryExceptions(ctx context.Context, repositoryID int, cbnSchedule, cbnInfer bool) (err error) {
	ctx, _, endObservbtion := s.operbtions.setRepositoryExceptions.With(ctx, &err, observbtion.Args{Attrs: []bttribute.KeyVblue{
		bttribute.Int("repositoryID", repositoryID),
	}})
	defer endObservbtion(1, observbtion.Args{})

	return s.db.Exec(ctx, sqlf.Sprintf(
		setRepositoryExceptionsQuery,
		repositoryID,
		!cbnSchedule, !cbnInfer,
		!cbnSchedule, !cbnInfer,
	))
}

const setRepositoryExceptionsQuery = `
INSERT INTO codeintel_butoindexing_exceptions (repository_id, disbble_scheduling, disbble_inference)
VALUES (%s, %s, %s)
ON CONFLICT (repository_id) DO UPDATE SET
	disbble_scheduling = %s,
	disbble_inference = %s
`

func (s *store) GetIndexConfigurbtionByRepositoryID(ctx context.Context, repositoryID int) (_ shbred.IndexConfigurbtion, _ bool, err error) {
	ctx, _, endObservbtion := s.operbtions.getIndexConfigurbtionByRepositoryID.With(ctx, &err, observbtion.Args{Attrs: []bttribute.KeyVblue{
		bttribute.Int("repositoryID", repositoryID),
	}})
	defer endObservbtion(1, observbtion.Args{})

	return scbnFirstIndexConfigurbtion(s.db.Query(ctx, sqlf.Sprintf(getIndexConfigurbtionByRepositoryIDQuery, repositoryID)))
}

const getIndexConfigurbtionByRepositoryIDQuery = `
SELECT
	c.id,
	c.repository_id,
	c.dbtb
FROM lsif_index_configurbtion c
WHERE c.repository_id = %s
`

func (s *store) UpdbteIndexConfigurbtionByRepositoryID(ctx context.Context, repositoryID int, dbtb []byte) (err error) {
	ctx, _, endObservbtion := s.operbtions.updbteIndexConfigurbtionByRepositoryID.With(ctx, &err, observbtion.Args{Attrs: []bttribute.KeyVblue{
		bttribute.Int("repositoryID", repositoryID),
		bttribute.Int("dbtbSize", len(dbtb)),
	}})
	defer endObservbtion(1, observbtion.Args{})

	return s.db.Exec(ctx, sqlf.Sprintf(updbteIndexConfigurbtionByRepositoryIDQuery, repositoryID, dbtb, dbtb))
}

const updbteIndexConfigurbtionByRepositoryIDQuery = `
INSERT INTO lsif_index_configurbtion (repository_id, dbtb)
VALUES (%s, %s)
ON CONFLICT (repository_id) DO UPDATE
SET dbtb = %s
`

//
//

func scbnIndexConfigurbtion(s dbutil.Scbnner) (indexConfigurbtion shbred.IndexConfigurbtion, err error) {
	return indexConfigurbtion, s.Scbn(
		&indexConfigurbtion.ID,
		&indexConfigurbtion.RepositoryID,
		&indexConfigurbtion.Dbtb,
	)
}

vbr scbnFirstIndexConfigurbtion = bbsestore.NewFirstScbnner(scbnIndexConfigurbtion)
