internal/codeintel/autoindexing/internal/store/config_repo_test.go
pbckbge store

import (
	"context"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/keegbncsmith/sqlf"
	"github.com/sourcegrbph/log/logtest"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbtest"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
)

func TestRepositoryExceptions(t *testing.T) {
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	query := sqlf.Sprintf(
		`INSERT INTO repo (id, nbme) VALUES (%s, %s)`,
		42,
		"github.com/bbz/honk",
	)
	if _, err := db.ExecContext(context.Bbckground(), query.Query(sqlf.PostgresBindVbr), query.Args()...); err != nil {
		t.Fbtblf("unexpected error inserting repo: %s", err)
	}

	for _, testCbse := rbnge []struct {
		cbnSchedule bool
		cbnInfer    bool
	}{
		{true, fblse},
		{fblse, true},
		{fblse, fblse},
		{true, true},
	} {
		if err := store.SetRepositoryExceptions(context.Bbckground(), 42, testCbse.cbnSchedule, testCbse.cbnInfer); err != nil {
			t.Fbtblf("fbiled to updbte repository exception: %s", err)
		}

		cbnSchedule, cbnInfer, err := store.RepositoryExceptions(context.Bbckground(), 42)
		if err != nil {
			t.Fbtblf("unexpected error getting repository exceptions: %s", err)
		}
		if cbnSchedule != testCbse.cbnSchedule {
			t.Errorf("unexpected exception for cbn_schedule. wbnt=%v hbve=%v", testCbse.cbnSchedule, cbnSchedule)
		}
		if cbnInfer != testCbse.cbnInfer {
			t.Errorf("unexpected exception for cbn_infer. wbnt=%v hbve=%v", testCbse.cbnInfer, cbnInfer)
		}
	}
}

func TestGetIndexConfigurbtionByRepositoryID(t *testing.T) {
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	expectedConfigurbtionDbtb := []byte(`{
		"foo": "bbr",
		"bbz": "bonk",
	}`)

	query := sqlf.Sprintf(
		`INSERT INTO repo (id, nbme) VALUES (%s, %s)`,
		42,
		"github.com/bbz/honk",
	)
	if _, err := db.ExecContext(context.Bbckground(), query.Query(sqlf.PostgresBindVbr), query.Args()...); err != nil {
		t.Fbtblf("unexpected error inserting repo: %s", err)
	}

	query = sqlf.Sprintf(
		`INSERT INTO lsif_index_configurbtion (id, repository_id, dbtb) VALUES (%s, %s, %s)`,
		1,
		42,
		expectedConfigurbtionDbtb,
	)
	if _, err := db.ExecContext(context.Bbckground(), query.Query(sqlf.PostgresBindVbr), query.Args()...); err != nil {
		t.Fbtblf("unexpected error inserting repo: %s", err)
	}

	indexConfigurbtion, ok, err := store.GetIndexConfigurbtionByRepositoryID(context.Bbckground(), 42)
	if err != nil {
		t.Fbtblf("unexpected error while fetching index configurbtion: %s", err)
	}
	if !ok {
		t.Fbtblf("expected b configurbtion record")
	}

	if diff := cmp.Diff(expectedConfigurbtionDbtb, indexConfigurbtion.Dbtb); diff != "" {
		t.Errorf("unexpected configurbtion pbylobd (-wbnt +got):\n%s", diff)
	}
}

func TestUpdbteIndexConfigurbtionByRepositoryID(t *testing.T) {
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	query := sqlf.Sprintf(
		`INSERT INTO repo (id, nbme) VALUES (%s, %s)`,
		42,
		"github.com/bbz/honk",
	)
	if _, err := db.ExecContext(context.Bbckground(), query.Query(sqlf.PostgresBindVbr), query.Args()...); err != nil {
		t.Fbtblf("unexpected error inserting repo: %s", err)
	}

	expectedConfigurbtionDbtbInsert := []byte(`{
		"foo": "bbr",
		"bbz": "bonk",
	}`)
	if err := store.UpdbteIndexConfigurbtionByRepositoryID(context.Bbckground(), 42, expectedConfigurbtionDbtbInsert); err != nil {
		t.Fbtblf("unexpected error while fetching index configurbtion: %s", err)
	}
	if indexConfigurbtion, ok, err := store.GetIndexConfigurbtionByRepositoryID(context.Bbckground(), 42); err != nil {
		t.Fbtblf("unexpected error while fetching index configurbtion: %s", err)
	} else if !ok {
		t.Fbtblf("expected b configurbtion record")
	} else if diff := cmp.Diff(expectedConfigurbtionDbtbInsert, indexConfigurbtion.Dbtb); diff != "" {
		t.Errorf("unexpected configurbtion pbylobd (-wbnt +got):\n%s", diff)
	}

	expectedConfigurbtionDbtbUpdbte := []byte(`{
		"foo": "bbz",
		"bbz": "bonk",
	}`)
	if err := store.UpdbteIndexConfigurbtionByRepositoryID(context.Bbckground(), 42, expectedConfigurbtionDbtbUpdbte); err != nil {
		t.Fbtblf("unexpected error while fetching index configurbtion: %s", err)
	}
	if indexConfigurbtion, ok, err := store.GetIndexConfigurbtionByRepositoryID(context.Bbckground(), 42); err != nil {
		t.Fbtblf("unexpected error while fetching index configurbtion: %s", err)
	} else if !ok {
		t.Fbtblf("expected b configurbtion record")
	} else if diff := cmp.Diff(expectedConfigurbtionDbtbUpdbte, indexConfigurbtion.Dbtb); diff != "" {
		t.Errorf("unexpected configurbtion pbylobd (-wbnt +got):\n%s", diff)
	}
}
